airbyte ci help finally works :D
