+++ airbyte-ci/connectors/pipelines/pipelines/pipeline/connectors/bump_version/pipeline.py
@@ -4,7 +4,7 @@ import datetime
 import semver
 from pipelines import consts
 from pipelines.helpers.connectors import metadata_change_helpers
-from pipelines.models.bases import Step, StepResult, StepStatus
+from pipelines.models.steps import Step, StepResult, StepStatus
 from pipelines.pipeline.connectors.context import ConnectorContext
 from pipelines.pipeline.connectors.reports import ConnectorReport
