+++ airbyte-ci/connectors/pipelines/pipelines/pipeline/connectors/migrate_to_base_image/pipeline.py
@@ -10,7 +10,7 @@ from connector_ops.utils import ConnectorLanguage
 from dagger import Directory
 from jinja2 import Template
 from pipelines import consts
-from pipelines.models.bases import Step, StepResult, StepStatus
+from pipelines.models.steps import Step, StepResult, StepStatus
 from pipelines.pipeline.connectors.bump_version.pipeline import BumpDockerImageTagInMetadata
 from pipelines.pipeline.connectors.bump_version.pipeline import AddChangelogEntry, get_bumped_version
 from pipelines.pipeline.connectors.context import ConnectorContext, PipelineContext
