+++ airbyte-ci/connectors/pipelines/pipelines/pipeline/metadata/pipeline.py
@@ -3,17 +3,21 @@ import uuid
 from typing import Optional
 
 import dagger
-from pipelines.dagger.containers.python import with_python_base, with_pip_packages
-from pipelines.models.bases import Report, Step, StepResult
+from pipelines.dagger.actions.python.common import with_pip_packages
+from pipelines.dagger.containers.python import with_python_base
+from pipelines.models.steps import Step, StepResult
+from pipelines.models.reports import Report
 from pipelines.consts import DOCS_DIRECTORY_ROOT_PATH
 from pipelines.pipeline.connectors.context import ConnectorContext, PipelineContext
 from pipelines.helpers.steps import run_steps
-from pipelines.models.steps import PoetryRunStep, MountPath, SimpleDockerStep
+from pipelines.models.steps import MountPath
 from pipelines.helpers.utils import DAGGER_CONFIG, get_secret_host_variable
 from pipelines.consts import DOCS_DIRECTORY_ROOT_PATH
 from pipelines.pipeline.connectors.context import ConnectorContext
-from pipelines.models.steps import MountPath, SimpleDockerStep
-from pipelines.internal_tools import INTERNAL_TOOL_PATHS
+from pipelines.models.steps import MountPath
+from pipelines.internal_tools.internal import INTERNAL_TOOL_PATHS
+from pipelines.pipeline.steps.docker import SimpleDockerStep
+from pipelines.pipeline.steps.poetry import PoetryRunStep
 
 # STEPS
