+++ airbyte-ci/connectors/pipelines/tests/test_builds/test_python_connectors.py
@@ -5,7 +5,7 @@
 from pathlib import Path
 
 import pytest
-from pipelines.models.bases import StepStatus
+from pipelines.models.steps import StepStatus
 from pipelines.pipeline.connectors.builds.steps import python_connectors
 from pipelines.pipeline.connectors.builds.steps import build_customization
 from pipelines.pipeline.connectors.context import ConnectorContext
